Parser: Add EOF token type
This allows the parser to know when the file has ended and count
it as another token. This will only have a transition on the 'all'
type ones, meaning seeing if it should exit the node. Before it'd
just get stuck in a loop as the last token was repeated.